Modify: Area, Pref models go to accounts app
